refactor, fixed indexing, readme
