+++ README.md
@@ -3,4 +3,5 @@ Bot do Discorda do pobrania losowego słowa ze strony https://www.miejski.pl/
 
 Aby dodać bota na serwer właściciel serwera powinien wejść na link https://discord.com/api/oauth2/authorize?client_id=755066931136036988&permissions=2048&scope=bot, a następnie pozwolić botowi pisać na serwerze.
 
-Wywołanie bota następuje przez wpisanie komendy "!miejski"
+Wywołanie bota następuje przez wpisanie komendy "!miejski".
+Wywołanie statystyk użytkowników następuje poprzez wpisanie komendy "!stats".
